[ya-req] code separation: mkt, act, pay + ya-client changes moved to ext repo
